# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salon SaaS API",
        "description": "API for salon booking, appointments, staff, and billing management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "description": "Get list of all services, optionally filtered by category_id",
                "operationId": "getServices",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (0 or 1)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Service"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "description": "Create a new service",
                "operationId": "storeService",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "service_category_id",
                                    "title",
                                    "price",
                                    "image",
                                    "status"
                                ],
                                "properties": {
                                    "service_category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Hair Cut"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Professional hair cutting",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 30,
                                        "nullable": true
                                    },
                                    "image": {
                                        "description": "Service image file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "description": "0 = inactive, 1 = active",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "orderby": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/v1/services/{service}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "description": "Get a single service",
                "operationId": "getService",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "description": "Update a service. We use POST with _method=PUT to support multipart/form-data for image uploads in PHP.",
                "operationId": "updateService",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "status"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "Method spoofing for PUT",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "service_category_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "title": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "image": {
                                        "description": "Service image file (optional on update)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "0 = inactive, 1 = active",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "orderby": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "description": "Delete a service",
                "operationId": "deleteService",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Service ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Service deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Service": {
                "title": "Service",
                "description": "Service model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "service_category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Hair Cut"
                    },
                    "slug": {
                        "type": "string",
                        "example": "hair-cut"
                    },
                    "description": {
                        "type": "string",
                        "example": "Professional hair cutting",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "duration": {
                        "type": "integer",
                        "example": 30,
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "example": "http://localhost/storage/services/image.jpg"
                    },
                    "status": {
                        "description": "0=inactive, 1=active",
                        "type": "integer",
                        "example": 1
                    },
                    "orderby": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieAuth": {
                "type": "apiKey",
                "description": "Laravel Sanctum session cookie authentication.",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}